Move CLIs to CLI module
