+++ lib/launch_base.rb
@@ -1,2 +1,2 @@
-require 'launch_base/cli'
+require 'launch_base/cli/base'
 require 'launch_base/version'
