+++ lib/launch_base/cli/all.rb
@@ -1,0 +1,2 @@
+require_relative 'add'
+require_relative 'lint'
