+++ spec/launch_base/cli_spec.rb
@@ -1,8 +1,8 @@
-describe LaunchBase::CLI do
+describe LaunchBase::CLI::Base do
   describe 'new' do
     it 'generates a new app using the template' do
       stub_installed_rails_version('Rails 5.2.0')
-      expect_any_instance_of(LaunchBase::CLI).to receive(:run).with(/rails new foobar.+launch_base_default_template.rb/)
+      expect_any_instance_of(cli_base_class).to receive(:run).with(/rails new foobar.+launch_base_default_template.rb/)
 
       invoke_command 'new', 'foobar'
     end
@@ -57,7 +57,7 @@ describe LaunchBase::CLI do
 
   describe 'update' do
     it 'runs `bundle update launch_base`' do
-      expect_any_instance_of(LaunchBase::CLI).to receive(:system).with('bundle update launch_base --conservative')
+      expect_any_instance_of(cli_base_class).to receive(:system).with('bundle update launch_base --conservative')
 
       invoke_command 'update'
     end
@@ -80,7 +80,7 @@ describe LaunchBase::CLI do
 
   describe 'lint update' do
     it 'updates the gem and installs the linter configuration files' do
-      expect_any_instance_of(LaunchBase::CLI).to receive(:system).with('bundle update launch_base --conservative')
+      expect_any_instance_of(cli_base_class).to receive(:system).with('bundle update launch_base --conservative')
 
       within_temp_test_directory do
         invoke_command 'lint', 'update'
